fix(noUpdate constrain): change strategy to prevent a model attribute from being updated
I have a problem with the this.getDataValue method in the setter. When creating a resource it returns null or undefined
before setting the initial value. this.setDataValue works fine and I can get the resource with its expected values. But
when I want to update the resource later, this.getDataValues always returns undefined, so always accept the update.

To solve this problem, the only solution I found is to delegate the constraint to the service before calling the model.
 # Please enter the commit message for your changes. Lines starting